+++ z0bug_odoo/z0bug_odoo/testenv/testenv.py
@@ -1,12 +1,24 @@
 # -*- coding: utf-8 -*-
 """Test Environment v2.0.12
 
-Copy this file in tests directory of your module.
-Please copy the documentation testenv.rst file too in your module.
-The __init__.py must import testenv.
-Your python test file should have to contain some following example lines:
+You can locate the recent testenv.py in testenv directory of module
+https://github.com/zeroincombenze/tools/tree/master/z0bug_odoo/testenv
 
-::
+For full documentation visit:
+https://zeroincombenze-tools.readthedocs.io/en/latest/pypi_z0bug_odoo/index.html
+https://z0bug-odoo.readthedocs.io/en/latest/
+https://github.com/zeroincombenze/tools
+https://github.com/zeroincombenze/zerobug-test
+
+Copy the testenv.py file in tests directory of your module.
+Please copy the documentation testenv.rst file in your module too.
+
+The __init__.py must import testenv.
+
+    from . import testenv
+    from . import test_<MY_TEST_FILE>
+
+Your python test file have to contain some following example lines:
 
     import os
     import logging
@@ -14,7 +26,6 @@ Your python test file should have to contain some following example lines:
 
     _logger = logging.getLogger(__name__)
 
-    TEST_RES_PARTNER = {...}
     TEST_SETUP_LIST = ["res.partner", ]
 
     class MyTest(SingleTransactionCase):
@@ -23,18 +34,13 @@ Your python test file should have to contain some following example lines:
             super().setUp()
             # Add following statement just for get debug information
             self.debug_level = 2
-            data = {"TEST_SETUP_LIST": TEST_SETUP_LIST}
-            for resource in TEST_SETUP_LIST:
-                item = "TEST_%s" % resource.upper().replace(".", "_")
-                data[item] = globals()[item]
-            self.declare_all_data(data)                 # TestEnv swallows the data
-            self.setup_env()                            # Create test environment
+            self.setup_env()                # Create test environment
 
         def tearDown(self):
             super().tearDown()
-            if os.environ.get("ODOO_COMMIT_TEST", ""):  # pragma: no cover
+            if os.environ.get("ODOO_COMMIT_TEST", ""):
                 # Save test environment, so it is available to dump
-                self.env.cr.commit()                    # pylint: disable=invalid-commit
+                self.env.cr.commit()     # pylint: disable=invalid-commit
                 _logger.info("✨ Test data committed")
 
         def test_mytest(self):
@@ -43,67 +49,468 @@ Your python test file should have to contain some following example lines:
             )
             ...
 
-        def test_mywizard(self):
-            self.wizard(...)                # Test requires wizard simulator
+Model data declaration
+~~~~~~~~~~~~~~~~~~~~~~
 
-External reference
+Each model is declared in a csv file or xlsx file in test/data directory of the
+module. The file name is the same of model name with dots replaced by undescore.
+
+i.e. below the contents of res_parter.csv file:
+
+    id,name,street
+    z0bug.partner1,Alpha,"1, First Avenue"
+
+The model may also be declared in a dictionary which key which is the external
+reference used to retrieve the record. See online documentation for furthermore info.
+
+    Please, do not to declare ``product.product`` records: they are automatically
+    created as child of ``product.template``. The external reference must contain
+    the pattern ``_template`` (see below).
+
+Magic relationship
 ~~~~~~~~~~~~~~~~~~
 
-Every record is tagged by an external reference.
-The external reference may be:
+Some models/tables should be managed together, i.e. account.move and account.move.line.
+TestEnv manages these models/tables, called header/detail, just as a single object.
+When header record is created, all detail lines are created with header.
+Odoo standard declaration requires the details data in child reference field with
+command 0, 0.
+This method make unreadable the source data. Look at the simple follow example with
+usually Odoo declaration way:
 
-* Ordinary Odoo external reference (a), called xref, format "module.name"
-* Test reference, format "z0bug.name" (b)
-* Key value, format "external.key" (c)
-* 2 keys reference, for header/detail relationship (d)
-* Magic reference for 'product.template' / 'product.product' (e)
-* Magic 3 level Odoo xref, format "module.name.field" (f)
+    sale_order_data = {
+        "example.order_1": {
+            "partner_id": self.env.ref("base.res_partner_1"),
+            "origin": "example",
+            ...
+            "order_line": [
+                (0, 0, {
+                    "product_id": self.env.ref("product.product_product_1"),
+                    "product_qty": 1,
+                    "price_unit": 1.23,}),
+                (0, 0, {
+                    "product_id": self.env.ref("product.product_product_2"),
+                    "product_qty": 2,
+                    "price_unit": 2.34,}),
+            ]
+        }
 
-Ordinary Odoo external reference (a) is a record of 'ir.model.data';
+    }
+
+Now look at the same data in internal declaration by **z0bug_odoo**:
+
+    TEST_SALE_ORDER = {
+        "example.order_1": {
+            "partner_id": "base.res_partner_1",
+            "origin": "example",
+            ...
+        }
+
+    }
+
+    TEST_SALE_ORDER_LINE = {
+        "example.order_1_1": {
+            "product_id": "product.product_product_1",
+            "product_qty": 1,
+            "price_unit": 1.23,
+        },
+        "example.order_1_2": {
+            "product_id": "product.product_product_2",
+            "product_qty": 2,
+            "price_unit": 2.34,
+        }
+    }
+
+As you can see, the data is easy readable and easy updatable. Please, notice:
+
+#. Sale order lines are declared in specific model sale.order.line
+#. Record ID **must** begin with header ID, followed by "_" and line ID
+#. Reference data do not require self.env.ref(): they are automatically referenced
+
+It is also easy write the csv or xlsx file. This is the example with above data
+
+    id,partner_id,origin
+    example.order_1,base.res_partner_1,example
+
+    id,product_id,product_qty,price_unit
+    example.order_1_1,product.product_product_1,1,1.23
+    example.order_1_2,product.product_product_2,2,2.34
+
+In your test file you must declare the following statement:
+
+    TEST_SETUP_LIST = ["sale.order", "sale.order.line"]
+
+Another magic relationship is the product.template (product) / product.product (variant)
+relationship.
+Whenever a product.template (product) record is created,
+Odoo automatically creates one variant (child) record for product.product.
+If your test module does not need to manage product variants you can avoid to declare
+product.product data even if this model is used in your test data.
+
+For example, you have to test **sale.order.line** which refers to product.product.
+You simply declare a **product.template** record with external reference
+uses "_template" magic text.
+
+    TEST_PRODUCT_TEMPLATE = {
+        "z0bug.product_template_1": {
+            "name": "Product alpha",
+            ...
+        }
+    )
+
+    TEST_SALE_ORDER_LINE = {
+        "z0bug.order_1_1": {
+            "product_id": "z0bug.product_product_1",
+            ...
+        }
+    )
+
+Module test execution session
+-----------------------------
+
+Module test execution workflow should be:
+
+    #. Data declaration, in file .csv or .xlszìx or in source code
+    #. Base data creation, in setUp() function
+    #. Tests execution
+    #. Supplemental data creation, during test execution, by group name
+
+Test data may be managed by one or more data group; if not declared,
+"base" group name is used. The "base" group will be created at the setUp()
+level: it is the base test data.
+Testing function may declare and manage other group data. Look at the
+following example:
+
+    import os
+    import logging
+    from .testenv import MainTest as SingleTransactionCase
+
+    _logger = logging.getLogger(__name__)
+
+    TEST_PRODUCT_TEMPLATE = {
+        "z0bug.product_template_1": {...}
+    }
+    TEST_RES_PARTNER = {
+        "z0bug.partner1": {...}
+    )
+    TEST_SETUP_LIST = ["res.partner", "product.template"]
+
+    TEST_SALE_ORDER = {
+        "z0bug.order_1": {
+            "partner_id": "z0bug.partner1",
+            ...
+        }
+    }
+    TEST_SALE_ORDER_LINE = {
+        "z0bug.order_1_1": {
+            "product_id": "z0bug.product_product_1",
+            ...
+        }
+    )
+
+    class MyTest(SingleTransactionCase):
+
+        def setUp(self):
+            super().setUp()
+            self.debug_level = 2
+            self.setup_env()                # Create base test environment
+
+        def test_something(self):
+            # Now add Sale Order data, group "order"
+            self.setup_env(group="order", setup_list=["sale.order", "sale.order.line"])
+
+Note the external reference are globals and they are visible from any groups.
+After base data is created, the real test session can begin. You can simulate
+various situation; the most common are:
+
+    #. Simulate web form create record
+    #. Simulate web form update record
+    #. Simulate the multi-record windows action
+    #. Download any binary data created by test
+    #. Engage wizard
+
+.. note::
+
+    You can also create / update record with usually create() / write() Odoo function,
+    but they do not really simulate the user behavior because they do not engage the
+    onchange methods, they do not load any view and so on.
+
+The real best way to test a create record is like the follow example
+based on res.partner model:
+
+        partner = self.resource_edit(
+            resource="res.partner",
+            web_changes=[
+                ("name", "Adam"),
+                ("country_id", "base.us"),
+                ...
+            ],
+        )
+
+You can also simulate the update session, issuing the record:
+
+        partner = self.resource_edit(
+            resource=partner,
+            web_changes=[
+                ("name", "Adam Prime"),
+                ...
+            ],
+        )
+
+Look at resource_edit() documentation for furthermore details.
+
+In you test session you should need to test a wizard. This test is very easy
+to execute as in the follow example that engage the standard language install
+wizard:
+
+        # We engage language translation wizard with "it_IT" language
+        # see "<ODOO_PATH>/addons/base/module/wizard/base_language_install*"
+        _logger.info("🎺 Testing wizard.lang_install()")
+        act_windows = self.wizard(
+            module="base",
+            action_name="action_view_base_language_install",
+            default={
+                "lang": "it_IT"
+                "overwrite": False,
+            },
+            button_name="lang_install",
+        )
+        self.assertTrue(
+            self.is_action(act_windows),
+            "No action returned by language install"
+        )
+        # Now we test the close message
+        self.wizard(
+            act_windows=act_windows
+        )
+        self.assertTrue(
+            self.env["res.lang"].search([("code", "=", "it_IT")]),
+            "No language %s loaded!" % "it_IT"
+        )
+
+Look at wizard() documentation for furthermore details.
+
+Data values
+-----------
+
+Data values may be raw data (string, number, dates, etc.) or external reference
+or some macro.
+You can declare data value on your own but you can discover th full test environment
+in https://github.com/zeroincombenze/zerobug-test/mk_test_env/ and get data
+from this environment.
+
+
+boolean
+~~~~~~~
+
+You can declare boolean value:
+
+* by python boolean False or True
+* by integer 0 or 1
+* by string "0" or "False" or "1" or "True"
+
+    self.resource_create(
+        "res.partner",
+        xref="z0bug.partner1",
+        values={
+             {
+                ...
+                "supplier": False,
+                "customer": "True",
+                "is_company": 1,
+            }
+        }
+    )
+
+char / text
+~~~~~~~~~~~
+
+Char and Text values are python string; please use unicode whenever is possible
+even when you test Odoo 10.0 or less.
+
+You can evalute the field value engaging a simple python expression inside tags like in
+following syntax:
+
+    "<?odoo EXPRESSION ?>"
+
+The expression may be a simple python expression with following functions:
+
+.. $include testenv_usafe_expr.csv
+
+::
+
+    self.resource_create(
+        "res.partner",
+        xref="z0bug.partner1",
+        values={
+             {
+                "name": "Alpha",
+                "street": "1, First Avenue"
+                # Name of Caserta city
+                "city": "<? base.state_it_ce.name ?>",
+                # Reference: 'year/123'
+                "ref": "<? compute_date('####-##-##')[0:4] + '/123' ?>",
+            }
+        }
+    )
+
+integer / float / monetary
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Integer, Floating and Monetary values are python integer or float.
+If numeric value is issued as string, it is internally converted
+as integer/float.
+
+
+date / datetime
+~~~~~~~~~~~~~~~
+
+Date and Datetime value are managed in special way.
+They are processed by ``compute_date()`` function (read below).
+You can issue a single value or a 2 values list, 1st is the date,
+2nd is the reference date.
+
+    self.resource_create(
+        "res.partner",
+        xref="z0bug.partner1",
+        values={
+             {
+                ...
+                "activity_date_deadline": "####-1>-##",    # Next month
+                "signup_expiration": "###>-##-##",         # Next year
+                "date": -1,                                # Yesterday
+                "last_time_entries_checked":
+                    [+2, another_date],                    # 2 days after another day
+                "message_last_post": "2023-06-26",         # Specific date, ISO format
+            }
+        }
+    )
+
+many2one
+~~~~~~~~
+
+You can issue an integer (if you know exactly the ID)
+or an external reference. Read above about external reference.
+
+    self.resource_create(
+        "res.partner",
+        xref="z0bug.partner1",
+        values={
+             {
+                ...
+                "country_id": "base.it",                   # Odoo external reference
+                "property_account_payable_id":
+                    "z0bug.customer_account",              # Test record
+                "title": "external.Mister"                 # Record with name=="Mister"
+            }
+        }
+    )
+
+
+one2many / many2many
+~~~~~~~~~~~~~~~~~~~~
+
+The one2many and many2many field may contains one or more ID;
+every ID use the same above many2one notation with external reference.
+Value may be a string (just 1 value) or a list.
+
+    self.resource_create(
+        "res.partner",
+        xref="z0bug.partner1",
+        values={
+             {
+                ...
+                "bank_ids":
+                    [
+                        "base.bank_partner_demo",
+                        "base_iban.bank_iban_china_export",
+                    ],
+                "category_id": "base.res_partner_category_0",
+            }
+        }
+    )
+
+    You can also use tha classic Odoo syntax with commands:
+    You can integrate classic Odoo syntax with **z0bug_odoo external** reference.
+
+binary
+~~~~~~
+
+Binary file are supplied with os file name. Test environment load file and
+get binary value. File must be located in **tests/data** directory.
+
+    self.resource_create(
+        "res.partner",
+        xref="z0bug.partner1",
+        values={
+             {
+                ...
+                "image": "z0bug.partner1.png"
+            }
+        }
+    )
+
+External reference for many2one, one2many and many2many fields
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Every record tagged by an external reference may be:
+
+    * Ordinary Odoo external reference ``(a)``, format "module.name"
+    * Test reference, format "z0bug.name" ``(b)``
+    * Key value, format "external.key" ``(c)``
+    * 2 keys reference, for header/detail relationship ``(d)``
+    * Magic reference for **product.template** / **product.product** ``(e)``
+
+Ordinary Odoo external reference ``(a)`` is a record of ir.model.data;
 you can see them from Odoo GUI interface.
 
-Test reference (b) are like external reference (a) but they are visible just in the
-test environment. They are identified by "z0bug." prefix module name.
+Test reference ``(b)`` are visible just in the test environment.
+They are identified by "z0bug." prefix module name.
 
-External key reference (c) is identified by "external." prefix followed by
-the key value used to retrieve the record. The field "code" or "name" are usually used
-to search record; for account.tax the "description" field is used.
+External key reference ``(c)`` is identified by "external." prefix followed by
+the key value used to retrieve the record.
+If key value is an integer it is the record "id".
+The field "code" or "name" are used to search record;
+for account.tax the "description" field is used.
 Please set self.debug_level = 2 (or more) to log these field keys.
 
-The 2 keys reference (d) needs to address child record inside header record
+The 2 keys reference ``(d)`` needs to address child record inside header record
 at 2 level model (header/detail) relationship.
-The key MUST BE the couple of header key plus "_" and plus line key (usually 'sequence'
-field); the header key is the same key of the parent record. The line key may be:
-
-* the sequence value (if present n model)
-* the most meaningful field value
-* an index value
-
-i.e. "z0bug.invoice_1_3" means: line with sequence 3 of 'account.invoice.line'
-which is child of record "z0bug.invoice_1" of 'account.invoice'.
-i.e.: "EUR.2023-06-26" should be the key for res.currency.rate where "EUR" is the header
-key (res.currency) and "2023-06-26" is the date of rate.
+The key MUST BE the same key of the parent record,
+plus "_", plus line identifier (usually sequence field).
+i.e. ``z0bug.move_1_3`` means: line with sequence ``3`` of account.move.line
+which is child of record ``z0bug.move_1`` of account.move
 Please set self.debug_level = 2 (or more) to log these relationships.
 
-For 'product.template' (product) you must use '_template' text in reference (e).
-TestEnv inherit 'product.product' (variant) external reference (read above
-'Magic relationship).
+For product.template (product) you must use '_template' text in reference ``(e)``.
+TestEnv inherit product.product (variant) external reference
+(read above "Magic relationship").
 
-The magic 3 level Odoo xref is a special reference applicable on all kind of data.
-The format is like standard odoo xref with a 3th level which is the field name.
-i.e. "base.partner_1.name" means the field name of the standard Odoo external reference
-"base.partner_1". Notice:
+Examples:
 
-* every field of xref may be used
-* TestEnv does not match the type of current field and xref field
-* External reference may one of above (a) (b) (c) (d) or (e)
+::
 
-For furthermore information, please:
+    TEST_ACCOUNT_ACCOUNT = {
+        "z0bug.customer_account": {
+            "code": "", ...
+        }
+        "z0bug.supplier_account": {
+            "code": "111100", ...
+        }
+    )
 
-* Read file testenv.rst in this directory (if supplied)
-* Visit https://zeroincombenze-tools.readthedocs.io
-* Visit https://github.com/zeroincombenze/tools
-* Visit https://github.com/zeroincombenze/zerobug-test
+    ...
+
+    self.resource_edit(
+        partner,
+        web_changes = [
+            ("country_id", "base.it"),       # Odoo external reference (type a)
+            ("property_account_receivable_id",
+             "z0bug.customer_account"),      # Test reference (type b)
+            ("property_account_payable_id",
+             "external.111100"),             # External key (type c)
+        ],
+    )
 """
 from __future__ import unicode_literals
 
@@ -115,6 +522,7 @@ import os
 import re
 import sys
 from datetime import datetime, date
+import random
 
 from future.utils import PY2, PY3
 from past.builtins import basestring, long
@@ -259,13 +667,25 @@ class MainTest(test_common.TransactionCase):
         self.convey_record = {}
         if not hasattr(self, "assert_counter"):
             self.assert_counter = 0
+        self.module = None
         for item in self.__module__.split("."):
             if item not in ("odoo", "openerp", "addons"):
-                self.module = self.env["ir.module.module"].search(
+                modules = self.env["ir.module.module"].search(
                     [("name", "=", item)]
-                )[0]
-                if self.module:
+                )
+                if modules:
+                    self.module = modules[0]
                     break
+        self.z0bug_lib = z0bug_odoo_lib.Z0bugOdoo()
+        # self.set_datadir(os.path.join(os.path.dirname(__file__), "data"),
+        #                  raise_if_not_found=False)
+        self.set_datadir(get_module_resource(self.module.name, "tests", "data"),
+                         raise_if_not_found=False)
+        self.params = {
+            "compute_date": self.compute_date,
+            "random": random.random,
+            "ref": self.env.ref,
+        }
 
     def tearDown(self):
         super(MainTest, self).tearDown()
@@ -520,6 +940,17 @@ class MainTest(test_common.TransactionCase):
     # ------------------------------
     # --  Hierarchical functions  --
     # ------------------------------
+
+    def set_datadir(self, data_dir, merge=False, raise_if_not_found=True):
+        if not merge:
+            self.source = "local"
+        self.z0bug_lib.declare_data_dir(
+            data_dir, merge=merge, raise_if_not_found=raise_if_not_found)
+        if data_dir:
+            self.data_dir = data_dir
+
+    def get_test_name(self, resource):
+        return "TEST_%s" % self.z0bug_lib.get_pymodel(resource).upper()
 
     def _search4parent(self, resource, parent_resource=None):
         if resource == "product.product":
@@ -778,12 +1209,17 @@ class MainTest(test_common.TransactionCase):
         if ftype not in ("text", "binary", "html"):
             value = self._get_conveyed_value(resource, field, value, fmt=fmt)
         if ((isinstance(value, str) or (sys.version_info[0] == 2
-                                        and isinstance(value, unicode)))
-                and value.startswith("<?odoo")
-                and value.endswith("?>")
-                and len(value.split(".")) == 3):
-            xref, field = [x.strip() for x in value[6: -2].rsplit(".", 1)]
-            value = self.resource_browse(xref=xref)[field]
+                                        and isinstance(value, unicode)))):
+            x = re.match(r"(<\? *odoo)(.*)(\?>)", value)
+            if x:
+                expr = x.groups()[1].strip()
+                if re.match(r"[\w]+\.[\w]+\.[\w]+$", expr):
+                    xref, field = [x.strip() for x in value[6: -2].rsplit(".", 1)]
+                    value = self.resource_browse(xref=xref)[field]
+                else:
+                    value = eval(expr, self.params)
+                    if ftype in ("char", "text", "selection"):
+                        value = str(value)
         if value is None or (
             isinstance(value, basestring)
             and (value in ("None", r"\N") or field == "id")
@@ -959,20 +1395,10 @@ class MainTest(test_common.TransactionCase):
     # Return base64 (binary data) or string (filename with len<=64)
 
     @api.model
-    def _get_binary_filename(self, xref, bin_types=None):
-        binary_root = get_module_resource(self.module.name, "tests", "data")
-        if not bin_types:
-            binary_file = os.path.join(binary_root, xref)
-            if os.path.isfile(binary_file):
-                return binary_file
-        bin_types = bin_types or ["png", "jpg", "xml"]
-        if not is_iterable(bin_types):
-            bin_types = [bin_types]  # pragma: no cover
-        for btype in bin_types:
-            binary_file = os.path.join(binary_root, "%s.%s" % (xref, btype))
-            if os.path.isfile(binary_file):
-                return binary_file
-        return False  # pragma: no cover
+    def _get_binary_filename(self, xref, bin_types=[]):
+        return self.z0bug_lib.get_data_filename(xref,
+                                                bin_types=bin_types,
+                                                raise_if_not_found=False)
 
     @api.model
     def _get_binary_contents(self, value):
@@ -1850,7 +2276,7 @@ class MainTest(test_common.TransactionCase):
         return self.env.user.company_id
 
     def compute_date(self, date, refdate=None):
-        """Compute date
+        """Compute date against reference date or today
 
         Args:
             date (date or string or integer): formula
@@ -1964,10 +2390,10 @@ class MainTest(test_common.TransactionCase):
         domain = build_domain(domain, ln, values)
         if not domain:                                               # pragma: no cover
             if raise_if_not_found:
-                self.raise_error("Invalid search keys %s for model %s"
-                                 % (self.skeys[resource], resource))
-            self._logger.info("⚠ Invalid search keys %s for model %s"
-                              % (self.skeys[resource], resource))
+                self.raise_error("No value %s supplied for search keys %s for model %s"
+                                 % (values, self.skeys[resource], resource))
+            self._logger.info("⚠ No value %s supplied for search keys %s for model %s"
+                              % (values, self.skeys[resource], resource))
             return False
         record = self.env[resource].search(domain, limit=3)
         if len(record) != 1 and parent_rec and isinstance(ln, (int, long)):
@@ -2136,7 +2562,8 @@ class MainTest(test_common.TransactionCase):
             )
         return record
 
-    def declare_resource_data(self, resource, data, name=None, group=None, merge=None):
+    def declare_resource_data(
+            self, resource, data, name=None, group=None, merge="local"):
         """Declare data to load on setup_env().
 
         Args:
@@ -2144,24 +2571,26 @@ class MainTest(test_common.TransactionCase):
             data (dict): record data
             name (str): label of dataset; default is resource name
             group (str): used to manager group data; default is "base"
-            merge (str): merge data with public data (currently just "zerobug")
+            merge (str): values are ("local"|"zerobug")
 
         Raises:
             TypeError: if invalid parameters issued
         """
-        if not isinstance(data, dict):  # pragma: no cover
+        if not isinstance(data, dict):                              # pragma: no cover
             self.raise_error("Dictionary expected")
-        if merge and merge != "zerobug":  # pragma: no cover
-            self.raise_error("Invalid merge value: please use 'zerobug'")
+        if merge not in ("local", "zerobug"):                       # pragma: no cover
+            self.raise_error("Invalid value %s ('zerobug' or 'local')" % merge)
+        if merge == "zerobug":
+            self.z0bug_lib.declare_data_dir(self.data_dir, merge=True)
+
         data = self.unicodes(data)
         for xref in list(sorted(data.keys())):
             if merge == "zerobug":
-                zerobug = z0bug_odoo_lib.Z0bugOdoo().get_test_values(resource, xref)
+                zerobug = self.z0bug_lib.get_test_values(resource, xref)
                 for field in list(zerobug.keys()):
                     if (
                         field not in data[xref]
-                        and zerobug[field]
-                        and zerobug[field] not in ("None", r"\N")
+                        and zerobug[field] not in (None, "None", r"\N")
                     ):
                         data[xref][field] = zerobug[field]
             tnxl_xref = self._get_conveyed_value(None, None, xref)
@@ -2174,7 +2603,7 @@ class MainTest(test_common.TransactionCase):
                 resource, xref, data[tnxl_xref], group=group, name=name
             )
 
-    def declare_all_data(self, message, group=None, merge=None):
+    def declare_all_data(self, message, group=None, merge="local", data_dir=None):
         """Declare all data to load on setup_env().
 
         Args:
@@ -2183,30 +2612,27 @@ class MainTest(test_common.TransactionCase):
                 TEST_* (dict): resource data; * is the uppercase resource name where
                                dot are replaced by "_"; (see declare_resource_data)
             group (str): used to manager group data; default is "base"
-            merge (str): merge data with public data; may be path to data dir or zerobug
+            merge (str): values are ("local"|"zerobug")
+            data_dir (str): data directory, default is "tests/data"
 
         Raises:
             TypeError: if invalid parameters issued
         """
         self.log_stack()
-        if not isinstance(message, dict):  # pragma: no cover
+        if not isinstance(message, dict):                           # pragma: no cover
             self.raise_error("Dictionary expected")
-        if "TEST_SETUP_LIST" not in message:  # pragma: no cover
+        if "TEST_SETUP_LIST" not in message:                        # pragma: no cover
             self.raise_error("Key TEST_SETUP_LIST not found")
         group = group or "base"
-        if not merge and os.path.isdir(os.path.join("tests", "data")):
-            z0bug_odoo_lib.Z0bugOdoo().declare_data_dir(
-                os.path.abspath(os.path.join("tests", "data")))
-        elif os.path.isdir(merge):
-            z0bug_odoo_lib.Z0bugOdoo().declare_data_dir(merge)
-            merge = None
+        if data_dir:
+            self.set_datadir(data_dir, merge=merge != "local")
         for resource in message["TEST_SETUP_LIST"]:
-            item = "TEST_%s" % resource.upper().replace(".", "_")
-            if item not in message:  # pragma: no cover
+            item = self.get_test_name(resource)
+            if item not in message:                                 # pragma: no cover
                 self.raise_error("Key %s not found" % item)
         for resource in message["TEST_SETUP_LIST"]:
             self.log_lvl_1(" 🐜 declare_all_data(%s,group=%s)" % (resource, group))
-            item = "TEST_%s" % resource.upper().replace(".", "_")
+            item = self.get_test_name(resource)
             self.declare_resource_data(
                 resource, message[item], group=group, merge=merge
             )
@@ -2219,6 +2645,7 @@ class MainTest(test_common.TransactionCase):
             resource (str): Odoo model name or name assigned, i.e. "res.partner"
             xref (str): external reference
             group (str): if supplied select specific group data; default is "base"
+            try_again (bool): engage conveyed value
 
         Returns:
             dictionary with data or empty dictionary
@@ -2326,6 +2753,8 @@ class MainTest(test_common.TransactionCase):
             company (obj): company to update; if not supplied a new company is created
             xref (str): external reference or alias for main company
             partner_xref (str): external reference or alias for main company partner
+            recv_xref (str): external reference or alias for receivable account
+            pay_xref (str): external reference or alias for payable account
             bnk1_xref (str): external reference or alias for 1st liquidity bank
             values (dict): company data to update immediately
             group (str): if supplied select specific group data; default is "base"
@@ -2398,31 +2827,67 @@ class MainTest(test_common.TransactionCase):
         lang=None,
         locale=None,
         group=None,
+        source="local",
+        setup_list=None,
+        data_dir=None,
     ):
         """Create all record from declared data.
         This function starts the test workflow creating the test environment.
-        Test data must be declared before engage this function with declare_all_data()
-        function (see above).
+        Test data must be declared before engage this function by file .csv or
+        file .xlsx or by source declaration TEST_<MODEL>.
         setup_env may be called more times with different group value.
         If it is called with the same group, it recreates the test environment with
         declared values; however this feature might do not work for some reason: i.e.
         if test creates a paid invoice, the setup_env() cannot unlink invoice.
         If you want to recreate the same test environment, assure the conditions for
         unlink of all created and tested records.
-        If you create more test environment with different group you can use all data,
-        even record created by different group.
-        In this way you can test a complex process the evolved scenario.
+        If you create more test environment with different group you can grow the data
+        during test execution with complex scenario.
+        In this way you can create functional tests not only regression tests.
 
         Args:
             lang (str): install & load specific language
             locale (str): install locale module with CoA; i.e l10n_it
             group (str): if supplied select specific group data; default is "base"
+            source (str): values are ("local"|"zerobug")
+            setup_list (list): list of Odoo modelS; if missed use TEST_SETUP_LIST
+            data_dir (str): data directory, default is "tests/data"
 
         Returns:
             None
         """
-        if "TEST_SETUP_LIST" in inspect.stack()[1][0].globals:
-            pass
+        def init_resource_data(resource, data):
+            item = self.get_test_name(resource)
+            if item in inspect.stack()[1][0].f_globals:
+                data[item] = inspect.stack()[1][0].f_globals[item]
+            elif self.data_dir:
+                data[item] = {}
+                for k in self.z0bug_lib.get_test_xrefs(resource):
+                    data[item][k] = {}
+            else:
+                self.raise_error("No data supplied for %s" % resource)
+
+        if data_dir:
+            self.set_datadir(data_dir, merge=source != "local")
+        if (
+                not self.setup_data
+                and "TEST_SETUP_LIST" in inspect.stack()[1][0].f_globals
+        ):
+            # No old API, load external data
+            data = {
+                "TEST_SETUP_LIST": inspect.stack()[1][0].f_globals["TEST_SETUP_LIST"]
+            }
+            for resource in data["TEST_SETUP_LIST"]:
+                init_resource_data(resource, data)
+            self.declare_all_data(data)
+        elif setup_list:
+            data = {
+                "TEST_SETUP_LIST": setup_list
+            }
+            for resource in setup_list:
+                init_resource_data(resource, data)
+            self.declare_all_data(data)
+        setup_list = setup_list or self.get_resource_list(group=group)
         self._logger.info(
             "🎺🎺🎺 Starting test v2.0.12 (debug_level=%s)" % (self.debug_level)
         )
@@ -2436,7 +2901,7 @@ class MainTest(test_common.TransactionCase):
         if lang:  # pragma: no cover
             self.install_language(lang)
         self._convert_test_data(group=group)
-        for resource in self.get_resource_list(group=group):
+        for resource in setup_list:
             resource_parent = self.parent_resource.get(resource)
             for xref in sorted(self.get_resource_data_list(resource, group=group)):
                 if resource_parent:
@@ -2553,7 +3018,7 @@ class MainTest(test_common.TransactionCase):
         """
         self.log_stack()
         if field not in record:  # pragma: no cover
-            raise ValueError("Field %s not found in %s" % (field, record._name))
+            self.raise_error("Field %s not found in %s" % (field, record._name))
         return base64.b64decode(getattr(record, field))
 
     @api.model
@@ -2956,8 +3421,7 @@ class MainTest(test_common.TransactionCase):
                 ctr_assertion += self.tmpl_validate_record(template, rec)
             return ctr_assertion
 
-        # if [key for key in template["_MATCH"]][0][1] == record:
-        if template["_MATCH"] and template["_MATCH"].keys()[0][1] == record:
+        if [key for key in template["_MATCH"]][0][1] == record:
             for field in template.keys():
                 if field in (childs_name, "id") or field.startswith("_"):
                     continue
@@ -3021,9 +3485,7 @@ class MainTest(test_common.TransactionCase):
         ctr_assertion = self.tmpl_validate_record(template, records)
         matches = []
         for tmpl in template:
-            if not tmpl["_MATCH"]:
-                self.raise_error("One template item matches twice!\n%s" % tmpl)
-            elif tmpl["_MATCH"] not in matches:
+            if tmpl["_MATCH"] not in matches:
                 matches.append(tmpl["_MATCH"])
                 ctr_assertion += 1
             else: